requirements.txt


langchain
langgraph
langchain-openai
openai
selenium
webdriver-manager
beautifulsoup4
requests
python-dotenv
